Read native file as bytes
