+++ pkg/filesutils/nativefiles/readandwrite_test.go
@@ -9,7 +9,7 @@ import (
 )
 
 func Test_ReadAndWriteAsString(t *testing.T) {
-	t.Run("example", func(t *testing.T) {
+	t.Run("happy path", func(t *testing.T) {
 		ctx := getCtx()
 		tmpPath, err := tempfiles.CreateTemporaryFile(ctx)
 		require.NoError(t, err)
@@ -22,5 +22,22 @@ func Test_ReadAndWriteAsString(t *testing.T) {
 		require.NoError(t, err)
 
 		require.EqualValues(t, "hello world", content)
+	})
+}
+
+func Test_ReadAndWriteAsBytes(t *testing.T) {
+	t.Run("happy path", func(t *testing.T) {
+		ctx := getCtx()
+		tmpPath, err := tempfiles.CreateTemporaryFile(ctx)
+		require.NoError(t, err)
+		defer nativefiles.Delete(ctx, tmpPath)
+
+		err = nativefiles.WriteBytes(ctx, tmpPath, []byte("hello world"))
+		require.NoError(t, err)
+
+		content, err := nativefiles.ReadAsBytes(ctx, tmpPath)
+		require.NoError(t, err)
+
+		require.EqualValues(t, []byte("hello world"), content)
 	})
 }
